# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ElectricityGrid API",
    "description": "Real-time and historical electricity grid data for Australian networks (NEM & WEM)",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
...
  },
  "paths": {
...
    "/v1/stats": {
      "get": {
        "summary": "Aggregated network statistics over a time range",
        "tags": [
          "Historical Data"
        ],
        "description": "Returns min/max statistics for demand, renewables percentage, and price. Note: 24h range is not supported — use /v1/history instead.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "NEM",
                "WEM"
              ],
              "default": "NEM"
            },
            "in": "query",
            "name": "network",
            "required": false,
            "description": "Network to query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "7d",
                "30d",
                "90d"
              ],
              "default": "7d"
            },
            "in": "query",
            "name": "range",
            "required": false,
            "description": "Time range (24h not supported for stats)"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "NSW1",
                "QLD1",
                "VIC1",
                "SA1",
                "TAS1",
                "WEM"
              ]
            },
            "in": "query",
            "name": "region",
            "required": false,
            "description": "Optional — filter to a specific region"
          }
        ],
        "responses": {
          "200": {
            "description": "Network statistics",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Network statistics",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "demand_mw": {
                          "type": "object",
                          "properties": {
                            "min": {
                              "type": "object",
                              "properties": {
                                "value": {
                                  "type": "number",
                                  "nullable": true
                                },
                                "timestamp": {
                                  "type": "string",
                                  "nullable": true
                                }
                              }
                            },
                            "max": {
                              "type": "object",
                              "properties": {
                                "value": {
                                  "type": "number",
                                  "nullable": true
                                },
                                "timestamp": {
                                  "type": "string",
                                  "nullable": true
                                }
                              }
                            }
                          }
                        },
                        "renewables_pct": {
                          "type": "object",
                          "properties": {
                            "min": {
                              "type": "object",
                              "properties": {
                                "value": {
                                  "type": "number",
                                  "nullable": true
                                },
                                "timestamp": {
                                  "type": "string",
                                  "nullable": true
                                }
                              }
                            },
                            "max": {
                              "type": "object",
                              "properties": {
                                "value": {
                                  "type": "number",
                                  "nullable": true
                                },
                                "timestamp": {
                                  "type": "string",
                                  "nullable": true
                                }
                              }
                            }
                          }
                        },
                        "price": {
                          "type": "object",
                          "properties": {
                            "min": {
                              "type": "object",
                              "properties": {
                                "value": {
                                  "type": "number",
                                  "nullable": true
                                },
                                "timestamp": {
                                  "type": "string",
                                  "nullable": true
                                }
                              }
                            },
                            "max": {
                              "type": "object",
                              "properties": {
                                "value": {
                                  "type": "number",
                                  "nullable": true
                                },
                                "timestamp": {
                                  "type": "string",
                                  "nullable": true
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query params",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Invalid query params",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/history": {
      "get": {
        "summary": "Time series history for a specific metric",
        "tags": [
          "Historical Data"
        ],
        "description": "Returns a time series array for the requested metric, aggregated at the specified interval over the given range. Optionally filter by region or fuel technology.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "generation_mw",
                "price",
                "emissions_volume",
                "emission_intensity",
                "demand_mw",
                "renewables_pct"
              ]
            },
            "in": "query",
            "name": "metric",
            "required": true,
            "description": "The metric to retrieve time series data for"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "5m",
                "1h",
                "1d"
              ],
              "default": "1h"
            },
            "in": "query",
            "name": "interval",
            "required": false,
            "description": "Aggregation interval"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "24h",
                "7d",
                "30d",
                "90d"
              ],
              "default": "7d"
            },
            "in": "query",
            "name": "range",
            "required": false,
            "description": "Time window to query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "NEM",
                "WEM"
              ],
              "default": "NEM"
            },
            "in": "query",
            "name": "network",
            "required": false,
            "description": "Network to query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "NSW1",
                "QLD1",
                "VIC1",
                "SA1",
                "TAS1",
                "WEM"
              ]
            },
            "in": "query",
            "name": "region",
            "required": false,
            "description": "Optional — filter to a specific region"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "fueltech",
            "required": false,
            "description": "Optional — filter by fuel technology e.g. solar, wind, coal, gas_ccgt"
          }
        ],
        "responses": {
          "200": {
            "description": "Time series data",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Time series data",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "metric": {
                          "type": "string",
                          "example": "price"
                        },
                        "interval": {
                          "type": "string",
                          "example": "1h"
                        },
                        "range": {
                          "type": "string",
                          "example": "7d"
                        },
                        "series": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "timestamp": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "value": {
                                "type": "number",
                                "nullable": true
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query params — metric is required",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Invalid query params — metric is required",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://ec2-54-226-204-58.compute-1.amazonaws.com:3000",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}